btm_sco_hfp_hal: Change get_packet_size() return type to size_t
The users of get_packet_size() (bluetooth::audio::sco::(wbs|swb)::init)
actually accept the packet size as size_t, and in kernel wbs_pkt_len is
also defined as an unsigned value. There's no reason to define its
return type as signed int.

Bug: 327547486
Tag: #floss
Test: mmm packages/modules/Bluetooth
Flag: EXEMPT, only type change and get_packet_size() always returns 0 on
      Android.

+++ system/stack/btm/btm_sco_hfp_hal_linux.cc
@@ -48,7 +48,7 @@ struct mgmt_bt_codec {
 
 typedef struct cached_codec_info {
   struct bt_codec inner;
-  uint8_t pkt_size;
+  size_t pkt_size;
 } cached_codec_info;
 
 std::vector<cached_codec_info> cached_codecs;
@@ -451,7 +451,7 @@ void set_codec_datapath(int codec_uuid) {
   }
 }
 
-int get_packet_size(int codec) {
+size_t get_packet_size(int codec) {
   for (const cached_codec_info& c : cached_codecs) {
     if (c.inner.codec == codec) {
       return c.pkt_size;
